spiceserver tests added
