+++ isis/src/base/apps/spiceserver/spiceserver.h
@@ -1,0 +1,11 @@
+#ifndef spiceserver_h
+#define spiceserver_h
+
+#include "Pvl.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void spiceserver(UserInterface &ui, Pvl *log = nullptr);
+}
+
+#endif
