+++ isis/tests/FunctionalTestsSpiceserver.cpp
@@ -1,0 +1,172 @@
+#include <QTextStream>
+#include <QStringList>
+#include <QTemporaryFile>
+#include <QByteArray>
+#include <QFile>
+#include <QVector>
+#include <QDomDocument>
+
+#include "spiceserver.h"
+#include "Fixtures.h"
+#include "Pvl.h"
+#include "PvlGroup.h"
+#include "TestUtilities.h"
+#include "TextFile.h"
+
+#include "gmock/gmock.h"
+
+using namespace Isis;
+
+static QString APP_XML = FileName("$ISISROOT/bin/xml/spiceserver.xml").expanded();
+
+class TestPayload : public TempTestingFiles {
+    protected:
+      
+      QString asciiPayloadPath; 
+      QString hexPayloadPath; 
+
+      void SetUp() override {    
+        TempTestingFiles::SetUp();
+        QByteArray asciiPayload = R"(
+            <input_label>
+  <isis_version>
+ 352e312e312e3020616c706861207c20323031322d30352d3231 
+  </isis_version>
+  <parameters>
+    <cksmithed value='no' />
+    <ckrecon value='yes' />
+    <ckpredicted value='no' />
+    <cknadir value='no' />
+    <spksmithed value='no' />
+    <spkrecon value='yes' />
+    <spkpredicted value='no' />
+    <shape value='system' />
+    <startpad time='0.0' />
+    <endpad time='0.0' />
+  </parameters>
+  <label>
+4f626a656374203d2049736973437562650a20204f626a656374203d20436f72650a202020205374617274427974652020203d2036353533370a20202020466f726d61742020202020203d2054696c650a2020202054696c6553616d706c6573203d203132380a2020202054696c654c696e65732020203d203132380a0a2020202047726f7570203d2044696d656e73696f6e730a20202020202053616d706c6573203d20313230340a2020202020204c696e65732020203d20313035360a20202020202042616e64732020203d20310a20202020456e645f47726f75700a0a2020202047726f7570203d20506978656c730a20202020202054797065202020202020203d205265616c0a202020202020427974654f7264657220203d204c73620a20202020202042617365202020202020203d20302e300a2020202020204d756c7469706c696572203d20312e300a20202020456e645f47726f75700a2020456e645f4f626a6563740a0a202047726f7570203d20496e737472756d656e740a20202020537061636563726166744e616d65202020202020203d2056494b494e475f4f5242495445525f320a20202020496e737472756d656e7449642020202020202020203d2056495355414c5f494d4147494e475f53554253595354454d5f43414d4552415f420a202020205461726765744e616d6520202020202020202020203d204d4152530a20202020537461727454696d652020202020202020202020203d20313937372d30372d30375431313a33323a34350a202020204578706f737572654475726174696f6e20202020203d20302e303333393430203c7365636f6e64733e0a2020202053706163656372616674436c6f636b436f756e74203d2035323932333833390a20202020466c6f6f644d6f64654964202020202020202020203d204f4e0a202020204761696e4d6f6465496420202020202020202020203d204c4f570a202020204f66667365744d6f646549642020202020202020203d204f4e0a2020456e645f47726f75700a0a202047726f7570203d20417263686976650a20202020446174615365744964202020202020203d20564f312f564f322d4d2d5649532d322d4544522d56322e300a2020202050726f647563744964202020202020203d203331394231380a202020204d6973736f6e50686173654e616d65203d20455854454e4445445f4d495353494f4e0a20202020496d6167654e756d62657220202020203d2035323932333833390a202020204f726269744e756d62657220202020203d203331390a2020456e645f47726f75700a0a202047726f7570203d2042616e6442696e0a2020202046696c7465724e616d65203d20434c4541520a2020202046696c74657249642020203d20340a2020456e645f47726f75700a0a202047726f7570203d204b65726e656c730a202020204e6169664672616d65436f6465202020202020202020202020203d202d33303030320a202020204c6561705365636f6e64202020202020202020202020202020203d2024626173652f6b65726e656c732f6c736b2f6e616966303030392e746c730a2020202054617267657441747469747564655368617065202020202020203d2024626173652f6b65726e656c732f70636b2f70636b30303030392e7470630a20202020546172676574506f736974696f6e2020202020202020202020203d20285461626c652c2024626173652f6b65726e656c732f73706b2f64653430352e627370290a20202020496e737472756d656e74506f696e74696e6720202020202020203d20285461626c652c202476696b696e67322f6b65726e656c732f636b2f766f325f736564725f636b322e62632c0a2020202020202020202020202020202020202020202020202020202020202020202476696b696e67322f6b65726e656c732f666b2f766f325f7631302e7466290a20202020496e737472756d656e74202020202020202020202020202020203d204e756c6c0a2020202053706163656372616674436c6f636b20202020202020202020203d20282476696b696e67322f6b65726e656c732f73636c6b2f766f325f666963742e7473632c0a2020202020202020202020202020202020202020202020202020202020202020202476696b696e67322f6b65726e656c732f73636c6b2f766f325f6673632e747363290a20202020496e737472756d656e74506f736974696f6e20202020202020203d20285461626c652c202476696b696e67322f6b65726e656c732f73706b2f766f325f72636f6e2e627370290a20202020496e737472756d656e74416464656e64756d20202020202020203d202476696b696e67322f6b65726e656c732f69616b2f76696b696e67416464656e64756d3030332e74690a2020202053686170654d6f64656c202020202020202020202020202020203d2024626173652f74657374446174612f4d617273506c616e65746172795261646975735f3830626f74746f6d2e632d0a202020202020202020202020202020202020202020202020202020202020202075620a20202020496e737472756d656e74506f736974696f6e5175616c697479203d205265636f6e73747275637465640a20202020496e737472756d656e74506f696e74696e675175616c697479203d205265636f6e73747275637465640a2020202043616d65726156657273696f6e202020202020202020202020203d20310a2020456e645f47726f75700a0a2020232043616c6962726174696f6e206571756174696f6e20696e2076696b63616c0a202023204449286c2c7329203d2028312e302f286578702a773129292a47286c2c73292a286761696e2a4452286c2c73292b4443286c2c73292b6f6666742b6f666663290a202023207769746820207731203d2077302a282864697374302a646973743029202f202864697374312a646973743129290a20202320616e6420206f666674286c2c7329203d20412a6c202b20422a6c2a6c202b20432a73202b20442a6c2a73202b20450a202047726f7570203d20526164696f6d657472790a202020206f666663202020202020203d20302e300a2020202065787020202020202020203d2033332e34340a202020206761696e202020202020203d20322e300a2020202044522020202020202020203d202f776f726b2f76696b696e675f6e616469722f663331396231382e636c6e2e6375620a2020202044432020202020202020203d202476696b696e67322f63616c6962726174696f6e2f6361366f66662e63616c2e6375620a2020202047202020202020202020203d202476696b696e67322f63616c6962726174696f6e2f636136636c722e63616c2e6375620a2020202077302020202020202020203d2039312e380a2020202077312020202020202020203d203132322e31323137353039353536340a2020202064697374302020202020203d203234333834303030302e300a2020202064697374312020202020203d203231313431323037302e32323639380a20202020312e302f6578702a7731203d20322e3434383732383931303738363634652d30340a2020202041636f65666620202020203d20312e3333333633333130313932353338652d30350a2020202042636f65666620202020203d202d312e3232353035363634323933393737652d30380a2020202043636f65666620202020203d20352e3739363735383239383038343637652d30370a2020202044636f65666620202020203d202d312e3132303038303030303736353136652d30380a2020202045636f65666620202020203d20342e363837353339333432373937310a2020456e645f47726f75700a0a202047726f7570203d20526573656175730a202020204c696e6520202020203d2028332e393838343938353634333337392c20362e363031363937343138373434372c20382e323638303834313632373532312c0a20202020202020202020202020202020392e373137373433313737303537362c2031312e3031333337343733343237382c2031322e3134373636313135343838352c0a2020202020202020202020202020202031332e3739383230383539303235322c2031352e3031323633343230373636322c2031362e3639393139353635323135352c0a2020202020202020202020202020202031382e3335373235353336373934312c2031392e39303531303835353830312c203133332e36313832343230393036332c0a202020202020202020202020202020203133342e39373734303139363239312c203133362e38303734353639353934362c203133372e393531373939393639392c0a202020202020202020202020202020203133392e31303933383832373737312c203134302e31343032313437373338362c203134312e323330373437383834332c0a202020202020202020202020202020203134322e38323630343733313133322c203134332e39363531353035313132342c203134352e37313637393937323433392c0a202020202020202020202020202020203134372e36303835333739313036312c203134382e32303732373934353533362c203236342e30373732363839393234332c0a202020202020202020202020202020203236362e30333231373638363039332c203236372e31343232313136323431332c203236372e39313736383035303030382c0a202020202020202020202020202020203236382e393435323534333836322c203236392e38383834303836303539312c203237302e38353436383433313839392c0a202020202020202020202020202020203237322e31393633303731313533362c203237332e38323137323736313435382c203237352e35373636353536303931352c0a202020202020202020202020202020203237372e37303230393536363835312c203339342e30323933333435313638362c203339352e303531393933333831382c0a202020202020202020202020202020203339362e32333637323431323233372c203339372e303832393530383738372c203339372e38393630303734383732372c0a202020202020202020202020202020203339382e38373135333530333637382c203339392e39343938343038393833382c203430302e38363131393337303237312c0a202020202020202020202020202020203430312e39323237313932313735362c203430332e32373733383531383535392c203430352e32373338353038363331332c0a202020202020202020202020202020203430362e38303131333836363335382c203532342e30303930303534313638322c203532352e31393635303937383238372c0a202020202020202020202020202020203532362e33323238323632343132312c203532362e38363032373030363331322c203532372e37373533343634323031342c0a202020202020202020202020202020203532372e393039303735393931332c203532382e37303032303337333536372c203533302e31343730373339373338322c0a202020202020202020202020202020203533312e31383437393833333835362c203533322e39353139313831343430322c203533352e33343035303133313639362c0a202020202020202020202020202020203635332e38373334393439363333312c203635342e33333632363230303938382c203635352e31343732383336393236342c0a202020202020202020202020202020203635352e37383833333730333138382c203635362e35393030353837313831372c203635362e39373538353830363837312c0a202020202020202020202020202020203635372e38363938313431303534322c203635382e31333734363136383532362c203635392e33383238323932373234382c0a202020202020202020202020202020203636302e38323739363934343034312c203636322e37303133393332303932392c203636332e38383233323234383030392c0a202020202020202020202020202020203738332e38303037323138353337332c203738342e32343836383738343037392c203738342e37333933363730303631372c0a202020202020202020202020202020203738352e32303436373932383239322c203738352e38343533303432323731372c203738352e39393832343833393830372c0a202020202020202020202020202020203738362e39383035343036303839392c203738372e37363730363135303737362c203738392e30353733373638393138382c0a202020202020202020202020202020203739302e31343833383738373135362c203739322e39333235303739393631372c203931332e32373438333030373933342c0a202020202020202020202020202020203931332e32393231363435353336312c203931332e38323833393435313437372c203931342e33303231353837303136342c0a202020202020202020202020202020203931342e37353133343139303138382c203931342e39373537383433333535362c203931352e36333331343931303532382c0a202020202020202020202020202020203931362e31333038343939323634352c203931362e39303432373830363839352c203931382e33303038383837343037332c0a202020202020202020202020202020203932302e31363934303832383130392c203932312e37353130363736383038332c20313034322e373032303737383632312c0a20202020202020202020202020202020313034322e373536303537303732382c20313034322e393934323832373637372c20313034332e323631373238363633332c0a20202020202020202020202020202020313034332e383632393538313437312c20313034332e393639313936303331382c20313034342e373031313334323631312c0a20202020202020202020202020202020313034352e323039313631363635372c20313034362e313532343138353338312c20313034382e303832363430393432322c0a20202020202020202020202020202020313035302e35323036303231323932290a2020202053616d706c652020203d202832312e3333373236393430353635352c203134302e32333037323830303434362c203235372e30373736343034333734342c0a202020202020202020202020202020203337322e38383734333139393633312c203438382e38333139383039353930322c203630342e31303339313332313034312c0a202020202020202020202020202020203731392e393131393538363334352c203833352e38393138373430323636312c203935312e39323331393133303131362c0a20202020202020202020202020202020313036382e333331393138373733382c20313138352e333331333238353732342c2032312e3632353736313131363735392c0a2020202020202020202020202020202038312e3733393335333035393639322c203139382e36353330333535383739352c203331352e30323232303031363038312c0a202020202020202020202020202020203433302e38393238303430303230352c203534362e36333733393636383037392c203636322e32353938393837313036312c0a202020202020202020202020202020203737372e39373031323734353835382c203839342e30323538393331343938372c20313031302e313531373335333930362c0a20202020202020202020202020202020313132362e323432393238333838352c20313138342e333231323439343630362c2032322e3233373435313431383337362c0a202020202020202020202020202020203134302e33383231383830353039332c203235372e30353035373435303537392c203337322e39383032343237383635322c0a202020202020202020202020202020203438382e39303536313731343338362c203630342e37363233393830323438332c203732302e35363434303232303636322c0a202020202020202020202020202020203833362e313938313135363936332c203935322e32333735363536303439392c20313036382e373131343030363334312c0a20202020202020202020202020202020313138342e373031323532373631322c2032322e3431353338383538383130382c2038312e3933393131313133343534332c0a202020202020202020202020202020203139392e32353539323837303233382c203331352e36303631363336333539322c203433312e31323135313335373934392c0a202020202020202020202020202020203534372e31343238393634393832372c203636332e31353938333233333733352c203737392e303836393134323836332c0a202020202020202020202020202020203839342e393631323139353636332c20313031312e333734333339343633392c20313132372e323237313235393634312c0a20202020202020202020202020202020313138352e363630383634363732392c2032322e3833353133383036343038312c203134302e38353230383632373832332c0a202020202020202020202020202020203235372e36363035343032393230392c203337332e37383133373631323737342c203438392e39343532303536333130342c0a202020202020202020202020202020203630352e37383930323337343131312c203732312e34303631353836303939362c203833372e33393333303431393236322c0a202020202020202020202020202020203935332e37363530333836313638322c20313036392e393130353630323435332c20313138352e393138363730313537372c0a2020202020202020202020202020202032322e3832313634363938313337392c2038322e3539373532313132303036312c203139392e35393534333432313333382c0a202020202020202020202020202020203331362e30323739323334353539362c203433322e30373935393837373637382c203534382e30353930353539323036362c0a202020202020202020202020202020203636332e39383536313338313937322c203738302e30303737323137393332332c203839362e313732373738323432382c0a20202020202020202020202020202020313031322e333339373033343332362c20313132382e3534353937303236312c20313138362e343834373138363234372c0a2020202020202020202020202020202032322e3631343237323739313939312c203134312e30373334373535383737382c203235372e36353430393730303433382c0a202020202020202020202020202020203337342e30333838373135353335322c203439302e31333538313033383436372c203630362e30333532373235303935362c0a202020202020202020202020202020203732322e31313133303430393938342c203833372e39363632363436313433332c203935342e36343639303535303531322c0a20202020202020202020202020202020313037302e393532383737383032332c20313138372e3537393936353438372c2032312e39363238323230383238332c0a2020202020202020202020202020202038312e3736333339323637323534392c203139392e30343433313035383837352c203331352e37383038383633373334342c0a202020202020202020202020202020203433312e39343535343931363230332c203534382e323439343835383334332c203636342e34343234343436343334342c0a202020202020202020202020202020203738302e32383134373735363135352c203839362e39333231323737393030362c20313031332e313731343139333731362c0a20202020202020202020202020202020313132392e333233373339363234312c20313138372e373336323133323237342c2032302e3335303138383731313237352c0a202020202020202020202020202020203133392e35373133333731343031352c203235362e37323431393132303634362c203337332e34353333393234323830352c0a202020202020202020202020202020203438392e38363436353831353638312c203630362e30363437343536373936342c203732312e39303432383630323139342c0a202020202020202020202020202020203833382e31383230323338383733372c203935342e37393237343436373839312c20313037312e343030323735373435332c0a20202020202020202020202020202020313138382e32353736383332303434290a202020205479706520202020203d2028312c20352c20352c20352c20352c20352c20352c20352c20352c20352c20362c20342c20352c20352c20352c20352c20352c20352c20352c20352c20352c0a20202020202020202020202020202020352c20362c20342c20352c20352c20352c20352c20352c20352c20352c20352c20352c20362c20342c20352c20352c20352c20352c20352c20352c20352c0a20202020202020202020202020202020352c20352c20352c20362c20342c20352c20352c20352c20352c20352c20352c20352c20352c20352c20362c20342c20352c20352c20352c20352c20352c0a20202020202020202020202020202020352c20352c20352c20352c20352c20362c20342c20352c20352c20352c20352c20352c20352c20352c20352c20352c20362c20342c20352c20352c20352c0a20202020202020202020202020202020352c20352c20352c20352c20352c20352c20352c20362c20342c20352c20352c20352c20352c20352c20352c20352c20352c20352c2036290a2020202056616c6964202020203d2028312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c0a20202020202020202020202020202020312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c0a20202020202020202020202020202020312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c0a20202020202020202020202020202020312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c0a20202020202020202020202020202020312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c20312c2031290a2020202054656d706c617465203d202476696b696e67322f726573656175732f766f322e766973622e74656d706c6174652e6375620a202020205374617475732020203d2052656d6f7665640a2020456e645f47726f75700a456e645f4f626a6563740a0a4f626a656374203d204c6162656c0a20204279746573203d2036353533360a456e645f4f626a6563740a0a4f626a656374203d205461626c650a20204e616d65202020202020202020202020202020203d20496e737472756d656e74506f696e74696e670a202053746172744279746520202020202020202020203d20353938313736300a202042797465732020202020202020202020202020203d2036340a20205265636f726473202020202020202020202020203d20310a2020427974654f7264657220202020202020202020203d204c73620a202054696d65446570656e64656e744672616d6573203d20282d33303030302c20322c2031290a2020436f6e7374616e744672616d65732020202020203d20282d33303030322c202d3330303030290a2020436f6e7374616e74526f746174696f6e202020203d2028302e303035383637393336303732353133372c20302e39393939313538383733363032342c0a202020202020202020202020202020202020202020202020202d302e3031313536363536393533363238352c202d302e393939393832343838363430322c0a20202020202020202020202020202020202020202020202020302e303035383538363539303030343431382c202d382e3335373739363831373439353738652d30342c0a202020202020202020202020202020202020202020202020202d372e3637393434373935333936323932652d30342c20302e3031313537313237313239313636382c0a20202020202020202020202020202020202020202020202020302e3939393933323735353730393835290a2020436b5461626c65537461727454696d65202020203d202d3730393630343738362e37353330380a2020436b5461626c65456e6454696d652020202020203d202d3730393630343738362e37353330380a2020436b5461626c654f726967696e616c53697a65203d20310a20204465736372697074696f6e2020202020202020203d202243726561746564206279207370696365696e6974220a20204b65726e656c73202020202020202020202020203d20282476696b696e67322f6b65726e656c732f636b2f766f325f736564725f636b322e62632c0a202020202020202020202020202020202020202020202020202476696b696e67322f6b65726e656c732f666b2f766f325f7631302e7466290a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303051300a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303051310a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303051320a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303051330a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204156310a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204156320a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204156330a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d2045540a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a456e645f4f626a6563740a0a4f626a656374203d205461626c650a20204e616d6520202020202020203d20496e737472756d656e74506f736974696f6e0a20205374617274427974652020203d20353938313832340a20204279746573202020202020203d2035360a20205265636f72647320202020203d20310a2020427974654f726465722020203d204c73620a20204361636865547970652020203d204c696e6561720a20204465736372697074696f6e203d202243726561746564206279207370696365696e6974220a20204b65726e656c7320202020203d202476696b696e67322f6b65726e656c732f73706b2f766f325f72636f6e2e6273700a0a202047726f7570203d204669656c640a202020204e616d65203d204a32303030580a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a32303030590a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a323030305a0a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303058560a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303059560a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a323030305a560a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d2045540a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a456e645f4f626a6563740a0a4f626a656374203d205461626c650a20204e616d65202020202020202020202020202020203d20426f6479526f746174696f6e0a202053746172744279746520202020202020202020203d20353938313838300a202042797465732020202020202020202020202020203d2036340a20205265636f726473202020202020202020202020203d20310a2020427974654f7264657220202020202020202020203d204c73620a202054696d65446570656e64656e744672616d6573203d202831303031342c2031290a2020436b5461626c65537461727454696d65202020203d202d3730393630343738362e37353330380a2020436b5461626c65456e6454696d652020202020203d202d3730393630343738362e37353330380a2020436b5461626c654f726967696e616c53697a65203d20310a20204465736372697074696f6e2020202020202020203d202243726561746564206279207370696365696e6974220a20204b65726e656c73202020202020202020202020203d202824626173652f6b65726e656c732f73706b2f64653430352e6273702c0a2020202020202020202020202020202020202020202020202024626173652f6b65726e656c732f70636b2f70636b30303030392e747063290a2020536f6c61724c6f6e6769747564652020202020203d203239332e33303831333631353632330a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303051300a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303051310a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303051320a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303051330a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204156310a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204156320a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204156330a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d2045540a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a456e645f4f626a6563740a0a4f626a656374203d205461626c650a20204e616d6520202020202020203d2053756e506f736974696f6e0a20205374617274427974652020203d20353938313934340a20204279746573202020202020203d2035360a20205265636f72647320202020203d20310a2020427974654f726465722020203d204c73620a20204361636865547970652020203d204c696e6561720a20204465736372697074696f6e203d202243726561746564206279207370696365696e6974220a20204b65726e656c7320202020203d2024626173652f6b65726e656c732f73706b2f64653430352e6273700a0a202047726f7570203d204669656c640a202020204e616d65203d204a32303030580a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a32303030590a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a323030305a0a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303058560a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a3230303059560a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d204a323030305a560a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a0a202047726f7570203d204669656c640a202020204e616d65203d2045540a2020202054797065203d20446f75626c650a2020202053697a65203d20310a2020456e645f47726f75700a456e645f4f626a6563740a0a4f626a656374203d20486973746f72790a20204e616d652020202020203d2049736973437562650a2020537461727442797465203d20353938323030300a2020427974657320202020203d2032373633360a456e645f4f626a6563740a0a4f626a656374203d204f726967696e616c4c6162656c0a20204e616d652020202020203d2049736973437562650a2020537461727442797465203d20353936333737370a2020427974657320202020203d20323037300a456e645f4f626a6563740a456e64
+  </label>
+</input_label>
+        )";
+
+        QByteArray hexPayload = asciiPayload.toHex(); 
+        
+        asciiPayloadPath = tempDir.path() + "/asciiPayload.txt";
+        hexPayloadPath = tempDir.path() + "/hexPayload.txt";
+        
+        QFile asciiFile(asciiPayloadPath);
+        asciiFile.open(QIODevice::WriteOnly);
+        asciiFile.write(asciiPayload);
+        asciiFile.close();
+        
+        QFile hexFile(hexPayloadPath);
+        hexFile.open(QIODevice::WriteOnly);
+        hexFile.write(hexPayload);
+        hexFile.close();
+    }
+};
+
+TEST_F(TestPayload, FunctionalTestSpiceserverDefaultParameters) {
+  QString outputFile = tempDir.path() + "out.txt";
+
+  QVector<QString> args = {"From="+hexPayloadPath, "To="+outputFile};
+  UserInterface options(APP_XML, args);
+  Pvl appLog;
+
+  spiceserver(options, &appLog);
+  
+  TextFile inFile(outputFile); 
+  QString hexCode; 
+  inFile.GetLine(hexCode);
+  
+  QString xml( QByteArray::fromHex( QByteArray( hexCode.toLatin1() ) ).constData() );
+  
+  QDomDocument document;
+  QString error;
+  Pvl kernelsLabel;
+  Pvl instrumentPositionTable;
+  
+  // Use Qt's XML API to get elements we want to compare 
+  int errorLine, errorCol;
+  if ( document.setContent(QString(xml), &error, &errorLine, &errorCol) ) {
+    QDomElement rootElement = document.firstChild().toElement();
+    for ( QDomNode node = rootElement.firstChild(); !node.isNull(); node = node.nextSibling() ) {
+      
+      QDomElement element = node.toElement();
+
+      if (element.tagName() == "kernels_label") {
+        QString encoded = element.firstChild().toText().data();
+        std::stringstream labStream;
+        labStream << QString( QByteArray::fromHex( encoded.toLatin1() ).constData() );
+        labStream >> kernelsLabel;
+      }
+      else if (element.tagName() == "tables") {
+        for ( QDomNode node = element.firstChild(); !node.isNull(); node = node.nextSibling() ) {
+          QDomElement table = node.toElement();
+          
+          if (table.tagName() == "instrument_position") {
+             QString encoded = table.firstChild().toText().data();
+             std::stringstream labStream;
+             labStream << QString( QByteArray::fromHex( encoded.toLatin1() ).constData() );
+             labStream >> instrumentPositionTable;
+          }
+        }
+      }
+    }
+  }
+  else {
+    FAIL() << "Unable to open output file, either doesn't exist for formatted incorrectly.";
+  }
+  
+  PvlGroup naifKeywords = kernelsLabel.group(0);
+
+  EXPECT_EQ((int)naifKeywords.findKeyword("NaifFrameCode"), -30002);
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, naifKeywords.findKeyword("TargetPosition")[0], "Table");
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, naifKeywords.findKeyword("InstrumentPointing")[0], "Table");
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, naifKeywords.findKeyword("InstrumentPosition")[0], "Table");
+  
+  PvlObject table = instrumentPositionTable.findObject("Table");
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, QString(table.findKeyword("Name")), "InstrumentPosition");
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, QString(table.group(0).findKeyword("Name")), "J2000X");
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, QString(table.group(1).findKeyword("Name")), "J2000Y");
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, QString(table.group(2).findKeyword("Name")), "J2000Z");
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, QString(table.group(3).findKeyword("Name")), "J2000XV");
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, QString(table.group(4).findKeyword("Name")), "J2000YV");
+  EXPECT_PRED_FORMAT2(AssertQStringsEqual, QString(table.group(5).findKeyword("Name")), "J2000ZV");
+}
+
+
+TEST_F(TempTestingFiles, FunctionalTestSpiceserverIsisVersion) {
+  
+  // Isis Version 3.4.1
+  QByteArray badPayload = R"(
+    <input_label>
+      <isis_version>
+    332e342e312e3020616c706861207c20323031322d30352d32310a
+      </isis_version>
+    </input_label>
+  )";
+  
+  QString outputFile = tempDir.path() + "out.txt"; 
+  QString badPayloadPath = tempDir.path() + "out.txt";
+  
+  QFile asciiFile(badPayloadPath);
+  asciiFile.open(QIODevice::WriteOnly);
+  asciiFile.write(badPayload.toHex());
+  asciiFile.close();
+
+  QVector<QString> args = {"From="+badPayloadPath, "To="+outputFile};
+  UserInterface options(APP_XML, args);
+  
+  EXPECT_THROW({
+    try{
+      spiceserver(options);
+      FAIL();
+    }
+    catch(const IException &e) {
+      EXPECT_THAT(e.what(), testing::HasSubstr("The SPICE server only supports Isis versions greater than or equal to 3.5.*.*"));
+      throw;
+    }
+  }, IException);
+}
